api/requirements.txt
# Web framework (choose one)
flask==2.0.1
# OR
# fastapi==0.68.0
# uvicorn==0.15.0

# PostgreSQL connector (equivalent to pg)
psycopg2-binary==2.9.1

# Additional useful packages
python-dotenv==0.19.0
requests==2.26.0